+++ funcs.py
@@ -1,4 +1,7 @@
 import re
+from bs4 import BeautifulSoup
+from datetime import datetime
+import httplib
 
 def isflairbotmessage(message):
   return message.subject == "account link replay"
@@ -12,6 +15,20 @@ def isMessageBodyValidLink(message):
     return 'http://'+matches.group(1)+'/replay'
   return False
 
+def bnetGet(region, url):
+  conn = httplib.HTTPConnection(region+'.battle.net')
+  conn.connect()
+  request = conn.putrequest('GET',url)
+  conn.endheaders()
+  conn.send('')
+  resp = conn.getresponse()
+  return resp.read()
+
+def getLeague(region, url):
+  print "getLeague(" + region[0] + ", " + url + ")"
+  soup = BeautifulSoup(bnetGet(region[0], "/sc2/"+region[1]+"/profile/"+url))
+  leaguename = soup.select(".badge-item")[0].select("span.badge")[0]['class'][1][6:]
+  return (leaguename, str(datetime.now().year), str(datetime.now().month), str(datetime.now().day))
 
 def readAccountsFile(fileName):
   def readAccountsFileLine(line):
